Bugfix.  Blocking on full queue in sub handler causes disconnect.
If a client-side subscription channel is full, the 'Handle' RPC
call will hang indefinitely, ultimately causing the connection to
drop.  We fix this by making subscription handlers non-blocking,
ie.:  subscription messages are dropped if the consumer is not
receiving them fast enough.  This is consistent with the behavior
of libp2p-pubsub.